@@ -30,8 +30,10 @@ class CrashController {
 
 	@GetMapping("/oups")
 	public String triggerException() {
-		throw new RuntimeException(
-				"Expected: controller used to showcase what " + "happens when an exception is thrown");
+		throw new RuntimeException("""
+				Expected: controller used to showcase what \
+				happens when an exception is thrown\
+				""");
 	}
 
 }
